Fixed some problems and added some stuff for goombas
-Now dies properly (and accordingly depending if its big notmario), only
problem is the squishing. It gives the wrong position value since it is
going lower than it's supposed to. the only time he dies improperly is
if he jumps on a goomba from a big height.
-when going from big notmario to normal notmario, time slows down for a
second and the player is invincible for 2 seconds(although he does not
go through goombas while invincible yet). Since I didnt know how to make
the player "flash" when he shrinks, he just turns blue while invincible.